requirements.txt
pillow
langchain_community
